`DimensionTab` to component
