+++ src/blocks/image-maxi/components/index.js
@@ -1,2 +1,2 @@
-// eslint-disable-next-line import/prefer-default-export
 export { default as HoverEffectControl } from './hover-effect-control';
+export { default as DimensionTab } from './dimension-tab';
